resolver auth verify

+++ leafx_API_getaway/src/resolvers/saleregister_resolvers.js
@@ -42,15 +42,65 @@ const saleRegisterResolvers = {
             }
         },
         saleRegisterById: async (_, {id}, {dataSources, userIdToken}) => {
-            return await dataSources.saleRegisterAPI.saleRegisterByIdRequest(id);// Para el delete, me toca pensar cómo cambiarlo
+            let saleData = await dataSources.saleRegisterAPI.saleRegisterByIdRequest(id);
+            id_cliente = saleData.cliente;
+            id_proveedor = saleData.proveedor;
+
+            let clienteData = await dataSources.clienteAPI.clienteByIdRequest(id_cliente);
+            let proveedorData = await dataSources.proveedorAPI.proveedorByIdRequest(id_proveedor);
+
+            user_id_cliente = clienteData.user;
+            user_id_proveedor = proveedorData.user;
+            
+            if(user_id_cliente == userIdToken || user_id_proveedor == userIdToken){
+                return saleData;
+            } else {
+                return null;
+            }
+                
         }
     },
     Mutation: {
-        createSaleRegister: async (_, { saleRegister }, {dataSources}) => {
-            return await dataSources.saleRegisterAPI.createSaleRegister(saleRegister);
+        createSaleRegister: async (_, { saleRegister }, {dataSources, userIdToken}) => {
+            let clienteData = await dataSources.clienteAPI.clienteByIdRequest(saleRegister.cliente);
+            let user_id =  clienteData.user;
+            
+            if(user_id == userIdToken){
+                let productData = await dataSources.productoAPI.productoByIdRequest(saleRegister.producto);
+                let id_proveedor = productData.proveedor;
+                let precio = productData.precio;
+
+                let saleData = {
+                    cliente: saleRegister.cliente,
+                    proveedor: id_proveedor,
+                    producto: saleRegister.producto,
+                    cantidad: saleRegister.cantidad,
+                    precio: precio,
+                    fecha: saleRegister.fecha
+                }
+
+                return await dataSources.saleRegisterAPI.createSaleRegister(saleData);
+            } else {
+                return null;
+            }
         },
-        deleteSaleRegister: async (_, { id }, {dataSources}) => {
-            return await dataSources.saleRegisterAPI.deleteSaleRegister(id);
+        deleteSaleRegister: async (_, { id }, {dataSources, userIdToken}) => {
+            let saleData = await dataSources.saleRegisterAPI.saleRegisterByIdRequest(id);
+            id_cliente = saleData.cliente;
+            id_proveedor = saleData.proveedor;
+
+            let clienteData = await dataSources.clienteAPI.clienteByIdRequest(id_cliente);
+            let proveedorData = await dataSources.proveedorAPI.proveedorByIdRequest(id_proveedor);
+
+            user_id_cliente = clienteData.user;
+            user_id_proveedor = proveedorData.user;
+            
+            if(user_id_cliente == userIdToken || user_id_proveedor == userIdToken){
+                await dataSources.saleRegisterAPI.deleteSaleRegister(id);
+                return "Registro de venta eliminado";
+            } else {
+                return null;
+            }            
         }
     }
 };
